src/services/producer/requirements.txt
Faker==30.1.0
fastavro==1.9.7
kafka-python==2.0.2
python-dateutil==2.9.0.post0
six==1.16.0
typing_extensions==4.12.2
